cs499-project-main/cs499-project-main/requirements.txt
numpy==1.24.2
resampy==0.4.2
soundfile==0.12.1
pytorch-lightning==2.0.2
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
transformers==4.28.1
python-Levenshtein==0.21.0
